# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tasvirnigor Film & Animation Studio API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  titleRu?: string | null;
  /** @nullable */
  titleTj?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionRu?: string | null;
  /** @nullable */
  descriptionTj?: string | null;
  youtubeUrl: string;
  bannerUrl: string;
  sortOrder: number;
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  titleRu?: string | null;
  /** @nullable */
  titleTj?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionRu?: string | null;
  /** @nullable */
  descriptionTj?: string | null;
  youtubeUrl: string;
  bannerUrl: string;
  sortOrder?: number;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  titleRu?: string | null;
  /** @nullable */
  titleTj?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionRu?: string | null;
  /** @nullable */
  descriptionTj?: string | null;
  youtubeUrl?: string;
  bannerUrl?: string;
  sortOrder?: number;
}

export interface TeamMember {
  id: number;
  name: string;
  position: string;
  /** @nullable */
  positionRu?: string | null;
  /** @nullable */
  positionTj?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  bioRu?: string | null;
  /** @nullable */
  bioTj?: string | null;
  photoUrl: string;
  sortOrder: number;
  createdAt: string;
}

export interface TeamMemberInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  position: string;
  /** @nullable */
  positionRu?: string | null;
  /** @nullable */
  positionTj?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  bioRu?: string | null;
  /** @nullable */
  bioTj?: string | null;
  photoUrl: string;
  sortOrder?: number;
}

export interface TeamMemberUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  position?: string;
  /** @nullable */
  positionRu?: string | null;
  /** @nullable */
  positionTj?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  bioRu?: string | null;
  /** @nullable */
  bioTj?: string | null;
  photoUrl?: string;
  sortOrder?: number;
}

export interface About {
  id: number;
  headline: string;
  /** @nullable */
  headlineRu?: string | null;
  /** @nullable */
  headlineTj?: string | null;
  body: string;
  /** @nullable */
  bodyRu?: string | null;
  /** @nullable */
  bodyTj?: string | null;
  /** @nullable */
  mission?: string | null;
  /** @nullable */
  missionRu?: string | null;
  /** @nullable */
  missionTj?: string | null;
  /** @nullable */
  founded?: string | null;
}

export interface AboutUpdate {
  headline?: string;
  /** @nullable */
  headlineRu?: string | null;
  /** @nullable */
  headlineTj?: string | null;
  body?: string;
  /** @nullable */
  bodyRu?: string | null;
  /** @nullable */
  bodyTj?: string | null;
  /** @nullable */
  mission?: string | null;
  /** @nullable */
  missionRu?: string | null;
  /** @nullable */
  missionTj?: string | null;
  /** @nullable */
  founded?: string | null;
}

export interface Contacts {
  id: number;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  telegram?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  youtube?: string | null;
  /** @nullable */
  facebook?: string | null;
}

export interface ContactsUpdate {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  telegram?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  youtube?: string | null;
  /** @nullable */
  facebook?: string | null;
}

export interface Partner {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionRu?: string | null;
  /** @nullable */
  descriptionTj?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  logoUrl: string;
  sortOrder: number;
  createdAt: string;
}

export interface PartnerInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionRu?: string | null;
  /** @nullable */
  descriptionTj?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  logoUrl: string;
  sortOrder?: number;
}

export interface PartnerUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionRu?: string | null;
  /** @nullable */
  descriptionTj?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  logoUrl?: string;
  sortOrder?: number;
}

export interface AdminCredentials {
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}