# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Tasvirnigor Film & Animation Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List all partners
      responses:
        "200":
          description: List of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create a partner (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
        "401":
          description: Unauthorized
  /partners/{id}:
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update a partner (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deletePartner
      tags: [partners]
      summary: Delete a partner (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
...
components:
  schemas:
...
    Partner:
      type: object
      required: [id, name, logoUrl, sortOrder, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        descriptionRu:
          type: ["string", "null"]
        descriptionTj:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        logoUrl:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
    PartnerInput:
      type: object
      required: [name, logoUrl]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        descriptionRu:
          type: ["string", "null"]
        descriptionTj:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        logoUrl:
          type: string
        sortOrder:
          type: integer
    PartnerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        descriptionRu:
          type: ["string", "null"]
        descriptionTj:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        logoUrl:
          type: string
        sortOrder:
          type: integer
...