# lib/api-zod/src/generated/types/partnerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tasvirnigor Film & Animation Studio API
 * OpenAPI spec version: 0.2.0
 */

export interface PartnerInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionRu?: string | null;
  /** @nullable */
  descriptionTj?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  logoUrl: string;
  sortOrder?: number;
}